feat(MoneyValidator): 로또 구입 금액 검사 로직을 validator에서 수행하도록 수정
